+++ src/api/config/server.config.ts
@@ -1,4 +1,4 @@
-export const Server = {
+export const ServerConfig = {
 	endpoint: process.env.NEXT_PUBLIC_APPWRITE_ENDPOINT ?? '',
 	projectId: process.env.NEXT_PUBLIC_APPWRITE_PROJECT_ID ?? '',
 	databaseId: process.env.APPWRITE_DATABASE_ID ?? '',
