+++ src/api/destinations.ts
@@ -1,12 +1,12 @@
 import {Destination} from '@/models/destination';
 
-import {database} from './appwrite';
-import {Server} from './server';
+import {database} from './config/appwrite.config';
+import {ServerConfig} from './config/server.config';
 
 export const getDestinations = async (): Promise<Destination[]> => {
 	const {documents: destinations} = await database.listDocuments<Destination>(
-		Server.databaseId,
-		Server.destinationCollectionId
+		ServerConfig.databaseId,
+		ServerConfig.destinationCollectionId
 	);
 
 	return destinations;
